feat(config,test) : generalize field validator to be used beyond EOSConfig
