+++ com.playeveryware.eos/Runtime/Core/Config/Attributes/FieldValidator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9209686f48242544997a71f68e50fa07
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
